Windows 64: Added an option to use a prebuilt plugin
In order to use this option, pass -DUSE_PREBUILT_PLUGIN:BOOL=ON
to configure line. This will skip compilation of 64 bit plugin
in the source tree and download prebuilt SLPlugin.exe and its
modules. This allows 64 bit viewer to use 32 bit prebuilt plugin
from the 32 bit viewer build making it possible to use parcel
media. Quicktime, 32 bit Flash are function in this way.